Avoid Settings app NPE on broken packages
Avoid Settings app NullPointerException that results from broken
uninstalled packages in a work profile or secondary user.

It's unclear what causes some packages to have a nonexistent
codePath/sourceDir, but in the event that they do, this shouldn't
cause Settings to crash. The NPE does not occur when reverting
I0ff770f2, but doing so is likely not desirable.

Issue: calyxos#1105
Change-Id: I7a1ec0f6ded2e93e29bec39146fd8007b77e2fc0
Signed-off-by: Rve27 <rve27github@gmail.com>